Add a Iter.Done function for determining if a cursor is still open
We would like to know if an Iter has been exhausted, or if it may
still have additional results. An Iter may have additional results
either if there are more results cached locally, or if we have a non-0
cursor ID. (When mongod returns the last batch of results, it returns
a cursor ID of 0).

Conceivably, you could determine this by calling Next and just storing
the result until you actually need it, but that's annoying. This is
simpler.

The semantics for tail cursors are a bit more complicated (since
conceivably you could call Done() and get false, but then call Next()
and get no results), but tail cursors were always complicated.